+++ lib/Models/Terria.ts
@@ -83,6 +83,7 @@ import updateModelFromJson from "./updateModelFromJson";
 import upsertModelFromJson from "./upsertModelFromJson";
 import ViewerMode from "./ViewerMode";
 import Workbench from "./Workbench";
+import IElementConfig from "./IElementConfig";
 // import overrides from "../Overrides/defaults.jsx";
 
 interface ConfigParameters {
@@ -267,6 +268,8 @@ export default class Terria {
   readonly catalog = new Catalog(this);
   readonly timelineClock = new Clock({ shouldAnimate: false });
   // readonly overrides: any = overrides; // TODO: add options.functionOverrides like in master
+
+  readonly elements = observable.map<string, IElementConfig>();
 
   @observable
   readonly mainViewer = new TerriaViewer(
@@ -977,6 +980,10 @@ export default class Terria {
 
     if (initData.catalog !== undefined) {
       this.catalog.group.addMembersFromJson(stratumId, initData.catalog);
+    }
+
+    if (isJsonObject(initData.elements)) {
+      this.elements.merge(initData.elements);
     }
 
     if (Array.isArray(initData.stories)) {