+++ CHANGES.md
@@ -7,6 +7,9 @@ Change Log
 
 * Fix `NotificationWindow` bug with `message`.
 * Re-add `loadInitSources` to `Terria.updateApplicationUrl()`
+* Added support for `elements` object in catalogue files (aka init files).
+  * Using this object you can hide/show most UI elements individually.
+  * See https://github.com/TerriaJS/terriajs/pull/5131. More in-depth docs to come.
 * [The next improvement]